+++ ExilenceNextApp/src/components/table-wrapper/TableWrapper.styles.ts
@@ -52,10 +52,9 @@ export const useStyles = makeStyles((theme: Theme) =>
       },
     },
     tableCell: {
-      padding: 16,
+      padding: 10,
       fontSize: '0.75rem',
       textAlign: 'left',
-      fontWeight: 300,
       lineHeight: 1.43,
       verticalAlign: 'inherit',
       color: theme.palette.text.primary,